+++ login/Makefile
@@ -2,8 +2,6 @@ LOGIN_DEPENDENCIES_TAG ?= "zitadel-login-dependencies:local"
 LOGIN_IMAGE_TAG ?= "zitadel-login:local"
 CORE_MOCK_TAG ?= "zitadel-core-mock:local"
 LOGIN_INTEGRATION_TESTSUITE_TAG ?= "zitadel-login-integration-testsuite:local"
-CORE_MOCK_CONTAINER_NAME ?= zitadel-mock-grpc-server
-LOGIN_CONTAINER_NAME ?= zitadel-login
 
 XDG_CACHE_HOME ?= $(HOME)/.cache
 export CACHE_DIR ?= $(XDG_CACHE_HOME)/zitadel-make
@@ -12,21 +10,18 @@ export CACHE_DIR ?= $(XDG_CACHE_HOME)/zitadel-make
 help:
 	@echo "Makefile for the login service"
 	@echo "Available targets:"
-	@echo "  help              	     - Show this help message"
-	@echo "  login                   - Start the login service"
-	@echo "  login-lint              - Run linting and formatting checks"
-	@echo "  login-lint-force        - Force run linting and formatting checks"
-	@echo "  login-unit              - Run unit tests"
-	@echo "  login-unit-force        - Force run unit tests"
-	@echo "  login-integration       - Run integration tests"
-	@echo "  login-integration-force - Force run integration tests"
-	@echo "  login-standalone        - Build the docker image for production login containers"
-	@echo "  login-quality           - Run all quality checks (login-lint, unit, integration)"
-	@echo "  login-ci                - Run all CI tasks. Run it with the -j flag to parallelize. make -j ci"
-	@echo "  show-cache-keys         - Show all cache keys with image ids and exit codes"
-	@echo "  clean-cache-keys        - Remove all cache keys"
-	@echo "  core-mock               - Start the core mock server"
-	@echo "  core-mock-stop          - Stop the core mock server"
+	@echo "  help              	     - Show this help message."
+	@echo "  login-lint              - Run linting and formatting checks. FORCE=true prevents skipping."
+	@echo "  login-lint-force        - Force run linting and formatting checks."
+	@echo "  login-unit              - Run unit tests. FORCE=true prevents skipping."
+	@echo "  login-unit-force        - Force run unit tests."
+	@echo "  login-integration       - Run integration tests. FORCE=true prevents skipping."
+	@echo "  login-integration-force - Force run integration tests."
+	@echo "  login-standalone-build  - Build the docker image for production login containers."
+	@echo "  login-quality           - Run all quality checks (login-lint, unit, integration)."
+	@echo "  login-ci                - Run all CI tasks. Run it with the -j flag to parallelize: make -j ci."
+	@echo "  show-cache-keys         - Show all cache keys with image ids and exit codes."
+	@echo "  clean-cache-keys        - Remove all cache keys."
 
 
 .PHONY: login-lint-force
@@ -47,13 +42,12 @@ login-unit:
 	./scripts/run_or_skip.sh login-unit-force $(LOGIN_DEPENDENCIES_TAG)
 
 .PHONY: login-integration-force
-login-integration-force: login core-mock login-integration-testsuite
-	docker run --rm $(LOGIN_INTEGRATION_TESTSUITE_TAG)
-	$(MAKE) core-mock-stop
+login-integration-force: login-standalone-build core-mock-build login-integration-testsuite-build
+	docker compose --file ./apps/login-integration-testsuite/docker-compose.yaml run --rm integration-testsuite
 
 .PHONY: login-integration
 login-integration:
-	./scripts/run_or_skip.sh login-integration-force '$(LOGIN_DEPENDENCIES_TAG);$(CORE_MOCK_TAG);$(LOGIN_INTEGRATION_TESTSUITE_TAG)'
+	./scripts/run_or_skip.sh login-integration-force '$(LOGIN_IMAGE_TAG);$(LOGIN_INTEGRATION_TESTSUITE_TAG);$(CORE_MOCK_TAG)'
 
 .PHONY: login-quality
 login-quality: core-mock-build login-quality-after-build
@@ -62,36 +56,21 @@ login-quality-after-build: login-lint login-unit login-integration
 
 .PHONY: login-ci
 login-ci: core-mock-build login-ci-after-build
-login-ci-after-build: login-quality-after-build login-standalone
+login-ci-after-build: login-quality-after-build login-standalone-build
 	@:
 
 login-dependencies:
 	docker buildx bake login-dependencies --set login-dependencies.tags=$(LOGIN_DEPENDENCIES_TAG);
 
-.PHONY: login-standalone
-login-standalone:
+.PHONY: login-standalone-build
+login-standalone-build:
 	docker buildx bake login-standalone --set login-standalone.tags=$(LOGIN_IMAGE_TAG);
-
-.PHONY: login
-login: login-standalone login-stop
-	docker run --detach --rm --name $(LOGIN_CONTAINER_NAME) --publish 3000:3000 $(LOGIN_IMAGE_TAG)
-
-login-stop:
-	docker rm --force $(LOGIN_CONTAINER_NAME) 2>/dev/null || true
 
 core-mock-build:
 	docker buildx bake core-mock --set core-mock.tags=$(CORE_MOCK_TAG);
 
-login-integration-testsuite: login-dependencies
+login-integration-testsuite-build: login-dependencies
 	docker buildx bake login-integration-testsuite --set login-integration-testsuite.tags=$(LOGIN_INTEGRATION_TESTSUITE_TAG)
-
-.PHONY: core-mock
-core-mock: core-mock-build core-mock-stop
-	docker run --detach --rm --name $(CORE_MOCK_CONTAINER_NAME) --publish 22221:22221 --publish 22222:22222 $(CORE_MOCK_TAG)
-
-.PHONY: core-mock-stop
-core-mock-stop:
-	docker rm --force $(CORE_MOCK_CONTAINER_NAME) 2>/dev/null || true
 
 .PHONY: clean-cache-keys
 clean-cache-keys:
