+++ login/apps/login-integration-testsuite/docker-compose.yaml
@@ -1,0 +1,31 @@
+services:
+  core-mock:
+    image: ${CORE_MOCK_TAG:-zitadel-core-mock:local}
+    container_name: core-mock
+    ports:
+      - 22220:22220
+      - 22222:22222
+
+  login:
+    image: ${LOGIN_IMAGE_TAG:-zitadel-login:local}
+    container_name: login
+    ports:
+      - 3000:3000
+    environment:
+      - ZITADEL_API_URL=http://core-mock:22222
+      - ZITADEL_SERVICE_USER_TOKEN="yolo"
+      - EMAIL_VERIFICATION=true
+      - DEBUG=true
+      - NEXT_PUBLIC_BASE_PATH=""
+
+  integration-testsuite:
+    image: ${LOGIN_INTEGRATION_TESTSUITE_TAG:-zitadel-login-integration-testsuite:local}
+    container_name: integration-testsuite
+    environment:
+      - LOGIN_BASE_URL=http://login:3000
+      - CYPRESS_CORE_MOCK_STUBS_URL=http://core-mock:22220/v1/stubs
+    depends_on:
+      login:
+        condition: service_started
+      core-mock:
+        condition: service_started
