+++ login/apps/login-integration-testsuite/integration/register-idp.cy.ts
@@ -1,4 +1,4 @@
-import { stub } from "../support/mock";
+import { stub } from "../support/e2e";
 
 const IDP_URL = "https://example.com/idp/url";
