+++ login/dockerfiles/login-base.Dockerfile
@@ -3,4 +3,4 @@ ENV PNPM_HOME="/pnpm"
 ENV PATH="$PNPM_HOME:$PATH"
 RUN corepack enable
 RUN apk add --no-cache libc6-compat bash git
-WORKDIR /app
+WORKDIR /build
